feat: add getWithResponse (#11)
~This PR is a breaking change~. This PR adds the method
`getWithResponse` to retrieve the `rawResponse`. This is necessary to
determine if a Portkey fallback happened or not.

Additionally, this PR adds another option for streaming which is
necessary to get usage data:
https://community.openai.com/t/usage-stats-now-available-when-using-streaming-with-the-chat-completions-api-or-completions-api/738156

EDIT: This is not a breaking change anymore, as `get` is just using
`getWithResponse`with directly selecting the data.